+++ src/everest/everest_storage.py
@@ -797,58 +797,6 @@ class EverestStorage:
 
         return batch_dataframes
 
-    def on_optimization_finished(self) -> None:
-        logger.debug("Storing final results Everest storage")
-
-        # This a somewhat arbitrary threshold, this should be a user choice
-        # during visualization:
-        CONSTRAINT_TOL = 1e-6
-
-        max_total_objective = -np.inf
-        for b in self.batches_with_function_results:
-            total_objective = b.batch_objectives["total_objective_value"].item()
-            bound_constraint_violation = (
-                0.0
-                if b.batch_bound_constraint_violations is None
-                else (
-                    b.batch_bound_constraint_violations.drop("batch_id")
-                    .to_numpy()
-                    .min()
-                    .item()
-                )
-            )
-            input_constraint_violation = (
-                0.0
-                if b.batch_input_constraint_violations is None
-                else (
-                    b.batch_input_constraint_violations.drop("batch_id")
-                    .to_numpy()
-                    .min()
-                    .item()
-                )
-            )
-            output_constraint_violation = (
-                0.0
-                if b.batch_output_constraint_violations is None
-                else (
-                    b.batch_output_constraint_violations.drop("batch_id")
-                    .to_numpy()
-                    .min()
-                    .item()
-                )
-            )
-            if (
-                max(
-                    bound_constraint_violation,
-                    input_constraint_violation,
-                    output_constraint_violation,
-                )
-                < CONSTRAINT_TOL
-                and total_objective > max_total_objective
-            ):
-                b.write_metadata(is_improvement=True)
-                max_total_objective = total_objective
-
     def export_dataframes(
         self,
     ) -> tuple[pl.DataFrame, pl.DataFrame, pl.DataFrame]:
